% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_guidelines.R
\name{check_guidelines}
\alias{check_guidelines}
\title{Check guidelines}
\usage{
check_guidelines(x = NULL)
}
\arguments{
\item{x}{A data.frame of the guidelines. IF NULL, pull from BC Data Catalogue.}
}
\value{
A invisible copy of the data.frame of the guidelines.
}
\description{
Check that updated DataBC guidelines are valid
}
