% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_guide.R
\name{mod_guide_ui}
\alias{mod_guide_ui}
\alias{mod_guide_server}
\title{mod_guide_ui and mod_guide_server}
\usage{
mod_guide_ui(id)

mod_guide_server(input, output, session)
}
\arguments{
\item{id}{shiny id}

\item{input}{internal}

\item{output}{internal}

\item{session}{internal}
}
\description{
A shiny Module.
}
\keyword{internal}
