% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkgtemplate-package.R
\docType{package}
\name{shinywqg-package}
\alias{shinywqg}
\alias{shinywqg-package}
\title{shinywqg: Water Quality Guideline Shiny App}
\description{
\if{html}{\figure{logo.png}{options: align='right'}}

User interface for the water quality guideline (WQG) website.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/poissonconsulting/shinywqg}
  \item Report bugs at \url{https://github.com/poissonconsulting/shinywqg/issues}
}

}
\author{
\strong{Maintainer}: Sebastian Dalgarno \email{seb@poissonconsulting.ca} (0000-0002-3658-4517)

}
\keyword{internal}
