% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shinywqg-package.R
\docType{package}
\name{shinywqg-package}
\alias{shinywqg}
\alias{shinywqg-package}
\title{shinywqg: Water Quality Guideline Shiny App}
\description{
\if{html}{\figure{logo.png}{options: align='right' alt='logo' width='120'}}

User interface to calculate water quality guidelines.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/poissonconsulting/shinywqg}
  \item Report bugs at \url{https://github.com/poissonconsulting/shinywqg/issues}
}

}
\author{
\strong{Maintainer}: Sebastian Dalgarno \email{seb@poissonconsulting.ca} (\href{https://orcid.org/0000-0002-3658-4517}{ORCID})

Other contributors:
\itemize{
  \item Joe Thorley \email{joe@poissonconsulting.ca} (\href{https://orcid.org/0000-0002-7683-4592}{ORCID}) [contributor]
  \item Province of British Columbia [copyright holder, funder]
}

}
\keyword{internal}
